# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Properties API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/auth/register": {
            "post": {
                "tags": [
                    "api authentication"
                ],
                "summary": "Register",
                "description": "Register user",
                "operationId": "v1-auth-register",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Eduard"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "email",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "email"
                        },
                        "example": "eduard.test@gmail.com"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "password",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "password"
                        },
                        "example": "abcdef"
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "description": "password confirmation",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "password"
                        },
                        "example": "abcdef"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "api authentication"
                ],
                "summary": "Login",
                "description": "Login user",
                "operationId": "v1-auth-login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "email",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "email"
                        },
                        "example": "eduard.test@gmail.com"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "password",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "password"
                        },
                        "example": "abcdef"
                    },
                    {
                        "name": "device_name",
                        "in": "query",
                        "description": "device name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "web"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/v1/auth/logout": {
            "get": {
                "tags": [
                    "api authentication"
                ],
                "summary": "Logout",
                "description": "Logout user",
                "operationId": "v1-auth-logout",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/v1/auth/user": {
            "get": {
                "tags": [
                    "api authentication"
                ],
                "summary": "User",
                "description": "get user",
                "operationId": "v1-auth-get-user",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email, password and token to get the authentication token",
                "name": "Token based auth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}